+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- (De)serialize `FixedPoint` as a string in case of human readable formats.
+
 ## [0.5.1] - 2021-05-28
 ### Added
 - docs: specify required features for feature-gated instances.